Moving starting items and inventory modifiers
Moved starting items, undroppable and character specific
